working version of k-WL

+++ Models/conv.py
@@ -2,11 +2,12 @@ import torch
 from torch_geometric.nn import MessagePassing
 import torch.nn.functional as F
 from torch_geometric.nn import global_mean_pool, global_add_pool
-from ogb.graphproppred.mol_encoder import AtomEncoder,BondEncoder
+from ogb.graphproppred.mol_encoder import AtomEncoder, BondEncoder
 from torch_geometric.utils import degree
 from torch_geometric.nn import GATv2Conv
 
 import math
+
 
 ### GAT
 # todo: implement GAT 
@@ -21,7 +22,7 @@ import math
 #         self.mlp = torch.nn.Sequential(torch.nn.Linear(emb_dim, 2*emb_dim), torch.nn.BatchNorm1d(2*emb_dim), torch.nn.ReLU(), torch.nn.Linear(2*emb_dim, emb_dim))
 #         self.eps = torch.nn.Parameter(torch.Tensor([0]))
 #         self.edge_encoder = edge_encoder
-            
+
 #     def forward(self, x, edge_index, edge_attr):
 #         edge_embedding = self.edge_encoder(edge_attr)
 #         out = self.mlp((1 + self.eps) *x + self.propagate(edge_index, x=x, edge_attr=edge_embedding))
@@ -35,7 +36,7 @@ import math
 
 #     def update(self, aggr_out):
 #         return aggr_out
-    
+
 ### GIN convolution along the graph structure
 class GINConv(MessagePassing):
     def __init__(self, emb_dim, edge_encoder):
@@ -43,25 +44,26 @@ class GINConv(MessagePassing):
             emb_dim (int): node embedding dimensionality
         '''
 
-        super(GINConv, self).__init__(aggr = "add")
+        super(GINConv, self).__init__(aggr="add")
 
-        self.mlp = torch.nn.Sequential(torch.nn.Linear(emb_dim, 2*emb_dim), torch.nn.BatchNorm1d(2*emb_dim), torch.nn.ReLU(), torch.nn.Linear(2*emb_dim, emb_dim))
+        self.mlp = torch.nn.Sequential(torch.nn.Linear(emb_dim, 2 * emb_dim), torch.nn.BatchNorm1d(2 * emb_dim),
+                                       torch.nn.ReLU(), torch.nn.Linear(2 * emb_dim, emb_dim))
         self.eps = torch.nn.Parameter(torch.Tensor([0]))
         self.edge_encoder = edge_encoder
-            
+
     def forward(self, x, edge_index, edge_attr):
         edge_embedding = self.edge_encoder(edge_attr)
-        out = self.mlp((1 + self.eps) *x + self.propagate(edge_index, x=x, edge_attr=edge_embedding))
+        out = self.mlp((1 + self.eps) * x + self.propagate(edge_index, x=x, edge_attr=edge_embedding))
 
         return out
 
     def message(self, x_j, edge_attr):
-
         # print(f"{x_j.shape}, {edge_attr.shape}")
         return F.relu(x_j + edge_attr)
 
     def update(self, aggr_out):
         return aggr_out
+
 
 ### GCN convolution along the graph structure
 class GCNConv(MessagePassing):
@@ -78,14 +80,15 @@ class GCNConv(MessagePassing):
 
         row, col = edge_index
 
-        #edge_weight = torch.ones((edge_index.size(1), ), device=edge_index.device)
-        deg = degree(row, x.size(0), dtype = x.dtype) + 1
+        # edge_weight = torch.ones((edge_index.size(1), ), device=edge_index.device)
+        deg = degree(row, x.size(0), dtype=x.dtype) + 1
         deg_inv_sqrt = deg.pow(-0.5)
         deg_inv_sqrt[deg_inv_sqrt == float('inf')] = 0
 
         norm = deg_inv_sqrt[row] * deg_inv_sqrt[col]
 
-        return self.propagate(edge_index, x=x, edge_attr = edge_embedding, norm=norm) + F.relu(x + self.root_emb.weight) * 1./deg.view(-1,1)
+        return self.propagate(edge_index, x=x, edge_attr=edge_embedding, norm=norm) + F.relu(
+            x + self.root_emb.weight) * 1. / deg.view(-1, 1)
 
     def message(self, x_j, edge_attr, norm):
         return norm.view(-1, 1) * F.relu(x_j + edge_attr)
@@ -100,7 +103,9 @@ class GNN_node(torch.nn.Module):
     Output:
         node representations
     """
-    def __init__(self, num_layer, emb_dim, drop_ratio = 0.5, JK = "last", residual = False, gnn_type = 'gin', node_encoder = lambda x: x, edge_encoder = lambda x: x):
+
+    def __init__(self, num_layer, emb_dim, drop_ratio=0.5, JK="last", residual=False, gnn_type='gin',
+                 node_encoder=lambda x: x, edge_encoder=lambda x: x):
         '''
             emb_dim (int): node embedding dimensionality
             num_layer (int): number of GNN message passing layers
@@ -130,8 +135,8 @@ class GNN_node(torch.nn.Module):
                 self.convs.append(GINConv(emb_dim, edge_encoder))
             elif gnn_type == 'gcn':
                 self.convs.append(GCNConv(emb_dim, edge_encoder))
-            elif gnn_type == 'gat': 
-                self.convs.append(GATv2Conv(in_channels = emb_dim, out_channels = emb_dime, edge_dim  = 1))
+            elif gnn_type == 'gat':
+                self.convs.append(GATv2Conv(in_channels=emb_dim, out_channels=emb_dime, edge_dim=1))
             else:
                 raise ValueError('Undefined GNN type called {}'.format(gnn_type))
 
@@ -139,7 +144,7 @@ class GNN_node(torch.nn.Module):
 
     def forward(self, batched_data):
         x, edge_index, edge_attr, batch = batched_data.x, batched_data.edge_index, batched_data.edge_attr, batched_data.batch
-         
+
         if self.gnn_type == "gat":
             edge_attr = self.edge_encoder(edge_attr)
         ### computing input node embedding
@@ -149,17 +154,17 @@ class GNN_node(torch.nn.Module):
             x = x.long()
         edge_attr = edge_attr.long()
 
-        h_list = [self.node_encoder(x)]     
+        h_list = [self.node_encoder(x)]
         for layer in range(self.num_layer):
 
             h = self.convs[layer](h_list[layer], edge_index, edge_attr)
             h = self.batch_norms[layer](h)
 
             if layer == self.num_layer - 1:
-                #remove relu for the last layer
-                h = F.dropout(h, self.drop_ratio, training = self.training)
+                # remove relu for the last layer
+                h = F.dropout(h, self.drop_ratio, training=self.training)
             else:
-                h = F.dropout(F.relu(h), self.drop_ratio, training = self.training)
+                h = F.dropout(F.relu(h), self.drop_ratio, training=self.training)
 
             if self.residual:
                 h += h_list[layer]
@@ -183,7 +188,9 @@ class GNN_node_Virtualnode(torch.nn.Module):
     Output:
         node representations
     """
-    def __init__(self, num_layer, emb_dim, drop_ratio = 0.5, JK = "last", residual = False, gnn_type = 'gin', node_encoder = lambda x: x, edge_encoder = lambda x: x, len_type = 0):
+
+    def __init__(self, num_layer, emb_dim, drop_ratio=0.5, JK="last", residual=False, gnn_type='gin',
+                 node_encoder=lambda x: x, edge_encoder=lambda x: x, len_type=0):
         '''
             emb_dim (int): node embedding dimensionality
         '''
@@ -221,29 +228,31 @@ class GNN_node_Virtualnode(torch.nn.Module):
                 self.convs.append(GINConv(emb_dim, edge_encoder))
             elif gnn_type == 'gcn':
                 self.convs.append(GCNConv(emb_dim, edge_encoder))
-            elif gnn_type == 'gat': 
-                self.convs.append(GATv2Conv(in_channels = emb_dim, out_channels = emb_dim, edge_dim = 1))
+            elif gnn_type == 'gat':
+                self.convs.append(GATv2Conv(in_channels=emb_dim, out_channels=emb_dim, edge_dim=1))
             else:
                 raise ValueError('Undefined GNN type called {}'.format(gnn_type))
 
             self.batch_norms.append(torch.nn.BatchNorm1d(emb_dim))
 
         for layer in range(num_layer - 1):
-            self.mlp_virtualnode_list.append(torch.nn.Sequential(torch.nn.Linear(emb_dim, 2*emb_dim), torch.nn.BatchNorm1d(2*emb_dim), torch.nn.ReLU(), \
-                                                    torch.nn.Linear(2*emb_dim, emb_dim), torch.nn.BatchNorm1d(emb_dim), torch.nn.ReLU()))
-
+            self.mlp_virtualnode_list.append(
+                torch.nn.Sequential(torch.nn.Linear(emb_dim, 2 * emb_dim), torch.nn.BatchNorm1d(2 * emb_dim),
+                                    torch.nn.ReLU(), \
+                                    torch.nn.Linear(2 * emb_dim, emb_dim), torch.nn.BatchNorm1d(emb_dim),
+                                    torch.nn.ReLU()))
 
     def forward(self, batched_data):
 
         x, edge_index, edge_attr, batch = batched_data.x, batched_data.edge_index, batched_data.edge_attr, batched_data.batch
-
 
         # Only works for integer features!!!
         x = x.long()
         edge_attr = edge_attr.long()
 
         ### virtual node embeddings for graphs
-        virtualnode_embedding = self.virtualnode_embedding(torch.zeros(batch[-1].item() + 1).to(edge_index.dtype).to(edge_index.device))
+        virtualnode_embedding = self.virtualnode_embedding(
+            torch.zeros(batch[-1].item() + 1).to(edge_index.dtype).to(edge_index.device))
         h_list = [self.node_encoder(x)]
         for layer in range(self.num_layer):
             ### add message from virtual nodes to graph nodes
@@ -254,10 +263,10 @@ class GNN_node_Virtualnode(torch.nn.Module):
 
             h = self.batch_norms[layer](h)
             if layer == self.num_layer - 1:
-                #remove relu for the last layer
-                h = F.dropout(h, self.drop_ratio, training = self.training)
+                # remove relu for the last layer
+                h = F.dropout(h, self.drop_ratio, training=self.training)
             else:
-                h = F.dropout(F.relu(h), self.drop_ratio, training = self.training)
+                h = F.dropout(F.relu(h), self.drop_ratio, training=self.training)
 
             if self.residual:
                 h = h + h_list[layer]
@@ -271,9 +280,12 @@ class GNN_node_Virtualnode(torch.nn.Module):
                 ### transform virtual nodes using MLP
 
                 if self.residual:
-                    virtualnode_embedding = virtualnode_embedding + F.dropout(self.mlp_virtualnode_list[layer](virtualnode_embedding_temp), self.drop_ratio, training = self.training)
+                    virtualnode_embedding = virtualnode_embedding + F.dropout(
+                        self.mlp_virtualnode_list[layer](virtualnode_embedding_temp), self.drop_ratio,
+                        training=self.training)
                 else:
-                    virtualnode_embedding = F.dropout(self.mlp_virtualnode_list[layer](virtualnode_embedding_temp), self.drop_ratio, training = self.training)
+                    virtualnode_embedding = F.dropout(self.mlp_virtualnode_list[layer](virtualnode_embedding_temp),
+                                                      self.drop_ratio, training=self.training)
 
         ### Different implementations of Jk-concat
         if self.JK == "last":
